Actual Fix FFMPG + rendered comments missing parts
So I accidentally uploaded the wrong file, haven't really slept or eaten for a while, sorry for that. Here's the description again tho: 

So, I found after some digging that the missing comments in the renders were due to regex, I believe I have fixed them after realizing it was the part after defining the max_chars, so I tested putting the "(\.|.$)" in https://regex101.com/ with the text from one of the best examples I had in my rendered results folder, then I altered it until the render was working. The closest I could get which seem to be perfect, but which might need more testing was: 
for x in re.finditer(rf" *((.{{0,{self.tts_module.max_chars}}})(.$| $| \n|\n)) 
Alternatively: 
for x in re.finditer(rf" *((.{{0,{self.tts_module.max_chars}}})(.$| $|\n | \n))

However, I'm still not sure if this is the best possible regex. Also, I added some new try: excepts: to solve the FFMPEG error, for some reason this just makes it work, but I would like to test it like 500 times for the same thread to see that it always gets the same amount of parts (and preferably the same total length). What I have done is to run main.py and then just let it fetch the temp .mp3 files, then restart when I see sufficient results, rather than render the whole .mp4 video. 

If anything this can inspire someone to improve the regex further to completely fix the missing parts in some rendered comments (in case this doesn't fully fix).